+++ test/panel_demo_scripts/test_totals_testing.gd
@@ -12,11 +12,21 @@ func test_generates_error():
 
 func test_generates_warning():
 	gut.get_logger().warn("This is a manual warning")
+	pass_test('passing')
 
 func test_multiple_passing_asserts():
 	assert_eq(1, 1)
 	assert_eq(2, 2)
 	assert_eq('a', 'a')
+	
+func test_multiple_failing():
+	assert_ne(1, 1)
+	assert_ne(2, 2)
+	assert_ne('a', 'a')
+	
+func test_multiple_failing2():
+	assert_ne(1, 1)
+	assert_ne(2, 2)
 
 func test_makes_orphan():
 	var orphan = Node2D.new()
@@ -28,5 +38,8 @@ func test_makes_an_info():
 
 func test_pending():
 	pending("this is pending")
+	
+func test_no_assert():
+	gut.p('this has no assert')