+++ test/unit/test_autofree.gd
@@ -2,6 +2,13 @@ extends 'res://addons/gut/test.gd'
 
 var AutoFree = load('res://addons/gut/autofree.gd')
 
+var _starting_orphans = 0
+func before_all():
+	_starting_orphans = gut.get_orphan_counter().orphan_count()
+	
+func after_all():
+	assert_eq(gut.get_orphan_counter().orphan_count(), _starting_orphans, 'no new orphans during script')
+
 func test_can_make_one():
 	assert_not_null(AutoFree.new())
 
@@ -65,7 +72,8 @@ func test_can_free_things_in_tree():
 
 func test_watch_for_orphans():
 	var n = autofree(Node.new())
-	assert_true(true)
+	pass_test('trying to make orphans')
 
 func test_watch_for_orphans2():
 	var n = autoqfree(Node.new())
+	pass_test('trying to make orphans')